Move contentpal interfaces to separete module. Implements #83
This allows to create a testing module without circular dependencies.
