+++ contactspal/src/main/java/org/dmfs/android/contactspal/rowsets/Dirty.java
@@ -20,13 +20,13 @@ import android.provider.ContactsContract;
 import android.support.annotation.NonNull;
 
 import org.dmfs.android.contactspal.tables.RawContacts;
+import org.dmfs.android.contentpal.ClosableIterator;
 import org.dmfs.android.contentpal.RowSet;
 import org.dmfs.android.contentpal.RowSnapshot;
 import org.dmfs.android.contentpal.View;
 import org.dmfs.android.contentpal.predicates.AnyOf;
 import org.dmfs.android.contentpal.predicates.EqArg;
 import org.dmfs.android.contentpal.rowsets.QueryRowSet;
-import org.dmfs.android.contentpal.tools.ClosableIterator;
 
 
 /**
