+++ contentpal-api/build.gradle
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2017 dmfs GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 apply plugin: 'com.android.library'
 
 android {
@@ -9,15 +25,8 @@ android {
         targetSdkVersion 25
         versionCode 1
         versionName "1.0"
+    }
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
     packagingOptions {
         exclude 'META-INF/DEPENDENCIES.txt'
         exclude 'META-INF/LICENSE.txt'
@@ -36,10 +45,4 @@ dependencies {
     compile 'com.android.support:support-annotations:' + SUPPORT_LIBRARY_VERSION
     compile 'org.dmfs:iterators:' + JEMS_VERSION
     compile 'org.dmfs:optional:' + JEMS_VERSION
-    compile project(':contentpal-api')
-
-    testCompile 'junit:junit:4.12'
-    testCompile 'org.robolectric:robolectric:3.1.4'
-    testCompile 'org.mockito:mockito-core:2.10.0'
-    testCompile 'org.dmfs:optional-testing:' + JEMS_VERSION
 }
