+++ contentpal-api/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<manifest package="org.dmfs.android.contentpal.api">
+
+    <application/>
+</manifest>
