+++ contentpal-api/src/main/java/org/dmfs/android/contentpal/InsertOperation.java
