+++ contentpal-api/src/main/java/org/dmfs/android/contentpal/OperationsBatch.java
