+++ contentpal-api/src/main/java/org/dmfs/android/contentpal/TransactionContext.java
