+++ contentpal/src/main/java/org/dmfs/android/contentpal/rowsets/AllRows.java
@@ -18,11 +18,11 @@ package org.dmfs.android.contentpal.rowsets;
 
 import android.support.annotation.NonNull;
 
+import org.dmfs.android.contentpal.ClosableIterator;
 import org.dmfs.android.contentpal.RowSet;
 import org.dmfs.android.contentpal.RowSnapshot;
 import org.dmfs.android.contentpal.View;
 import org.dmfs.android.contentpal.predicates.AnyOf;
-import org.dmfs.android.contentpal.tools.ClosableIterator;
 
 
 /**
