+++ contentpal/src/main/java/org/dmfs/android/contentpal/rowsets/Cached.java
@@ -18,9 +18,9 @@ package org.dmfs.android.contentpal.rowsets;
 
 import android.support.annotation.NonNull;
 
+import org.dmfs.android.contentpal.ClosableIterator;
 import org.dmfs.android.contentpal.RowSet;
 import org.dmfs.android.contentpal.RowSnapshot;
-import org.dmfs.android.contentpal.tools.ClosableIterator;
 import org.dmfs.android.contentpal.tools.FakeClosable;
 
 import java.io.IOException;
